Extract LogReader to a separate file
