+++ src/orquestra/sdk/_base/_logs/_interfaces.py
@@ -1,0 +1,44 @@
+################################################################################
+# © Copyright 2023 Zapata Computing Inc.
+################################################################################
+"""
+Logs-related interfaces.
+"""
+import typing as t
+
+from orquestra.sdk.schema.workflow_run import WorkflowRunId
+
+from orquestra.sdk.schema.ir import TaskInvocationId
+
+
+class LogReader(t.Protocol):
+    """
+    A component that reads logs produced by tasks and workflows.
+    """
+
+    def get_task_logs(
+        self, wf_run_id: WorkflowRunId, task_inv_id: TaskInvocationId
+    ) -> t.List[str]:
+        """
+        Reads all available logs, specific to a single task invocation/run.
+
+        Returns:
+            Log lines printed when running this task invocation. If the task didn't
+            produce any logs this should be an empty list.
+        """
+        ...
+
+    def get_workflow_logs(
+        self, wf_run_id: WorkflowRunId
+    ) -> t.Dict[TaskInvocationId, t.List[str]]:
+        """
+        Reads all available logs printed during execution of this workflow run.
+
+        Returns:
+            A mapping with task logs. Each key-value pair corresponds to one task
+            invocation.
+            - key: task invocation ID (see
+                orquestra.sdk._base.ir.WorkflowDef.task_invocations)
+            - value: log lines from running this task invocation
+        """
+        ...
