+++ src/orquestra/sdk/_ray/_dag.py
@@ -25,7 +25,8 @@ from .._base import _services, serde
 from .._base._db import WorkflowDB
 from .._base._env import RAY_GLOBAL_WF_RUN_ID_ENV
 from .._base._spaces._structs import ProjectRef
-from .._base.abc import LogReader, RuntimeInterface
+from .._base.abc import RuntimeInterface
+from .._base._logs._interfaces import LogReader
 from ..schema import ir
 from ..schema.configs import RuntimeConfiguration
 from ..schema.local_database import StoredWorkflowRun
